+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 **Bug Fixes**
 - When installing cyberware and selecting `None` for the Humanity Loss, the maximum humanity was not being decremented forcing one to re-calculate it using the calculator.
 - Corrected  an issue with Compendium Migration where `Scene` type Compendia was not properly being migrated.
+- #437 - Chat Card shows armor ablation even if target has no Armor
 # Version 0.81.2 (Hotfix) | Date: 2022-08-02
 **Bug Fixes**
 - #458 - The rulers for the DV Calculation is broken for tables with regexp characters in the name [such as (Autofire)]